requirements.txt
streamlit==1.22.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
pillow
